feat(useAPIForVADCDataDictionary): Added safe JSON parsing, moved table into its own folder, added console errors for server response data
